alpha-sniper/requirements.txt
ccxt
pandas
numpy
python-dotenv
requests
schedule
ta-lib
